// src/app/services/models/event-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { OrganiserDto } from '../models/organiser-dto';
import { UserDto } from '../models/user-dto';
export interface EventDto {
  createdDate?: string;
  eventDate?: string;
  eventDescription?: string;
  eventLocation?: string;
  eventName?: string;
  eventStatus?: 'ACTIVE' | 'COMPLETED' | 'CANCELLED' | 'FULL';
  eventTarget?: 'FAMILY' | 'SINGLES' | 'CHILDREN' | 'EVERYBODY' | 'ADULTS_ONLY';
  id?: number;
  maxAttendees?: number;
  organiser?: OrganiserDto;
  participants?: Array<UserDto>;
  ticketPrice?: number;
}